Responsive admin panel and refactor delete function

+++ components/ArtForm.jsx
@@ -1,7 +1,11 @@
-import { poppins } from "@/app/fonts";
 import React, { useEffect, useState } from "react";
+import DeleteIcon from "@mui/icons-material/Delete";
+import SaveAltIcon from "@mui/icons-material/SaveAlt";
+import { useRouter } from "next/navigation";
 
 function ArtForm({ art }) {
+  const router = useRouter();
+
   const [url, setUrl] = useState(art.url || "");
   const [name, setName] = useState(art.name || "");
   const [year, setYear] = useState(art.year || 0);
@@ -20,7 +24,8 @@ function ArtForm({ art }) {
   const [apiResponse, setApiResponse] = useState({});
   const [showDelete, setShowDelete] = useState(false);
 
-  // *** Fetch the
+  const apiPath = "http://localhost:5005/api";
+
   useEffect(() => {
     getPrices();
     let formats = [];
@@ -28,6 +33,9 @@ function ArtForm({ art }) {
     setToggleFormats(formats);
   }, [sale]);
 
+  // *** This function makes sure that checkboxes stay consistent with the input values.
+  // On the edit view they get pre-checked if a value is already on the DB entry.
+  // If a checkbox gets unchecked, the value is reset to 0 to avoid passing data to the create/edit function accidentally ***
   const handleCheckbox = (format, event) => {
     if (!toggleFormats.includes(format)) {
       setToggleFormats([...toggleFormats, format]);
@@ -42,6 +50,7 @@ function ArtForm({ art }) {
     }
   };
 
+  // *** This function takes the prices from the sale property of the entry and sets the states accordingly ***
   const getPrices = () => {
     sale.forEach((format) => {
       if (format.format === "original") setOriginalPrice(format.price);
@@ -50,6 +59,8 @@ function ArtForm({ art }) {
     });
   };
 
+  // *** This function will update the file image to the cloudinary folder and set the URL state
+  // with the returned value from the backend. ***
   const uploadImage = async (event) => {
     const formdata = new FormData();
     formdata.append("url", event.target.files[0]);
@@ -59,7 +70,7 @@ function ArtForm({ art }) {
       body: formdata,
     };
     try {
-      const response = await fetch("http://localhost:5005/api/upload", requestOptions);
+      const response = await fetch(`${apiPath}/upload`, requestOptions);
       const parsed = await response.json();
       setUrl(parsed.fileUrl);
     } catch (error) {
@@ -88,7 +99,7 @@ function ArtForm({ art }) {
       sale: saleList,
     };
 
-    const apiUrl = art._id ? `http://localhost:5005/api/art/item/${art._id}` : "http://localhost:5005/api/create";
+    const apiUrl = art._id ? `${apiPath}/art/item/${art._id}` : `${apiPath}/create`;
     const apiMethod = art._id ? "PATCH" : "POST";
 
     const response = await fetch(apiUrl, {
@@ -107,24 +118,30 @@ function ArtForm({ art }) {
     }
   };
 
-  const handleDelete = async (id, event) => {
-    event.preventDefault();
-
-    const response = await fetch(`http://localhost:5005/api/art/item/${id}`, {
-      method: "DELETE",
-    });
-    const parsed = response.json();
-    if (response.status === 200 || response.status === 201) {
-      setApiResponse({ status: "success", message: parsed.message });
-    }
-    if (response.status === 400) {
-      setApiResponse({ status: "error", message: response.message });
+  const handleDelete = async (id) => {
+    if (confirm("Are you sure you want to delete the item?")) {
+      try {
+        const response = await fetch(`${apiPath}/art/item/${id}`, {
+          method: "DELETE",
+        });
+        const parsed = response.json();
+        if (response.status === 200 || response.status === 201) {
+          setApiResponse({ status: "success", message: parsed.message });
+        } else if (response.status === 400) {
+          setApiResponse({ status: "error", message: response.message });
+        }
+      } catch (err) {
+        console.error("There was an error with the request: ", err);
+      }
+      router.push("/art");
+    } else {
+      return;
     }
   };
 
   return (
-    <>
-      <form className="flex-col m-auto md:w-1/2 mt-6" method="" action="">
+    <div className="w-full sm:w-3/4 md:w-1/2 xl:w-1/3 ">
+      <form className="flex-col mt-3" method="" action="">
         <label className="text-typo-600">
           URL
           <input className="border-2 p-1 rounded-sm w-full" type="file" onChange={(event) => uploadImage(event)} required />
@@ -176,31 +193,32 @@ function ArtForm({ art }) {
         </label>
         <hr className="mt-4 mb-2" />
         <h4>Size</h4>
-        <div className="grid grid-cols-11 gap-2 w-full items-center">
-          <label htmlFor="height">Height</label>
-          <input
-            id="height"
-            className="border-2 p-1 rounded-sm col-span-4"
-            type="number"
-            value={height}
-            onChange={(event) => setHeight(event.target.value)}
-            required
-          />
-          <span></span>
-          <label htmlFor="width">Width</label>
-          <input
-            id="width"
-            className="border-2 p-1 rounded-sm col-span-4"
-            type="number"
-            value={width}
-            onChange={(event) => setWidth(event.target.value)}
-            required
-          />
+        <div className="flex flex-col md:flex-row md:justify-between gap-1 w-full">
+          <label>
+            Height
+            <input
+              className="border-2 p-1 w-full rounded-sm"
+              type="number"
+              value={height}
+              onChange={(event) => setHeight(event.target.value)}
+              required
+            />
+          </label>
+          <label>
+            Width
+            <input
+              className="border-2 p-1 w-full rounded-sm"
+              type="number"
+              value={width}
+              onChange={(event) => setWidth(event.target.value)}
+              required
+            />
+          </label>
         </div>
         <hr className="mt-4 mb-2" />
         <h4>Sale</h4>
-        <label className="flex items-center gap-2 mb-1">
-          <div className="flex gap-2 items-center w-1/5">
+        <label className="flex w-full justify-between gap-2 mb-1">
+          <div className="flex w-1/3 gap-2 items-center">
             <input
               type="checkbox"
               checked={toggleFormats.includes("original")}
@@ -209,7 +227,7 @@ function ArtForm({ art }) {
             Original
           </div>
           <input
-            className="border-2 p-1 rounded-sm"
+            className="w-2/3 border-2 p-1 rounded-sm"
             type="number"
             placeholder="Price for original"
             disabled={!toggleFormats.includes("original")}
@@ -217,8 +235,8 @@ function ArtForm({ art }) {
             onChange={(event) => setOriginalPrice(event.target.value)}
           />
         </label>
-        <label className="flex gap-2 mb-1">
-          <div className="flex gap-2 items-center w-1/5">
+        <label className="flex w-full justify-between gap-2 mb-1">
+          <div className="flex w-1/3 gap-2 items-center">
             <input
               type="checkbox"
               checked={toggleFormats.includes("poster")}
@@ -227,7 +245,7 @@ function ArtForm({ art }) {
             Poster
           </div>
           <input
-            className="border-2 p-1 rounded-sm"
+            className="w-2/3 border-2 p-1 rounded-sm"
             type="number"
             placeholder="Price for poster"
             disabled={!toggleFormats.includes("poster")}
@@ -235,8 +253,8 @@ function ArtForm({ art }) {
             onChange={(event) => setPosterPrice(event.target.value)}
           />
         </label>
-        <label className="flex items-center gap-2">
-          <div className="flex gap-2 items-center w-1/5">
+        <label className="flex w-full justify-between gap-2 ">
+          <div className="flex w-1/3 gap-2 items-center">
             <input
               type="checkbox"
               checked={toggleFormats.includes("postcard")}
@@ -245,7 +263,7 @@ function ArtForm({ art }) {
             Postcard
           </div>
           <input
-            className="border-2 p-1 rounded-sm"
+            className="w-2/3 border-2 p-1 rounded-sm"
             type="number"
             placeholder="Price for postcard"
             disabled={!toggleFormats.includes("postcard")}
@@ -253,28 +271,29 @@ function ArtForm({ art }) {
             onChange={(event) => setPostCardPrice(event.target.value)}
           />
         </label>
-        <button className="bg-green-300 text-sm text-typo-900 rounded-lg mt-4 mx-auto p-2" type="submit" onClick={handleSubmit}>
-          {art._id ? "Edit Item" : "Create Item"}
-        </button>
-        {art._id && (
-          <button className="border-2 bg-red-300 rounded-lg mt-4 px-2 py-1" type="button" onClick={() => setShowDelete(true)}>
-            Delete Item
+        <div className="flex justify-start">
+          <button
+            className="flex items-center bg-green-600 text-sm text-white rounded-lg mt-4 p-2"
+            type="submit"
+            onClick={handleSubmit}
+          >
+            <SaveAltIcon className="mr-1" fontSize="small" />
+            {art._id ? "Edit Item" : "Create Item"}
           </button>
-        )}
-        {showDelete && (
-          <dialog>
-            <button type="button" onClick={() => setShowDelete(false)}>
-              Cancel
+          {art._id && (
+            <button
+              className="flex items-center bg-red-600 text-sm text-white rounded-lg mt-4 p-2 ml-2"
+              type="button"
+              onClick={() => handleDelete(art._id)}
+            >
+              <DeleteIcon className="mr-1" fontSize="small" /> Delete Item
             </button>
-            <button type="submit" onClick={(event) => handleDelete(art._id)}>
-              Cancel
-            </button>
-          </dialog>
-        )}
+          )}
+        </div>
       </form>
       {apiResponse.status === "success" && <div>{apiResponse.message}</div>}
       {apiResponse.status === "error" && <div>{apiResponse.message}</div>}
-    </>
+    </div>
   );
 }
